pyproject.toml
[project]
name = "polish-legal-docs-summarization-and-generation"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "bert-score>=0.3.13",
    "huggingface-hub>=0.34.4",
    "loguru>=0.7.3",
    "mltk>=0.0.5",
    "nltk>=3.9.1",
    "pdf2image>=1.17.0",
    "peft>=0.17.1",
    "pymupdf>=1.26.3",
    "pypdf>=6.0.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "rouge>=1.0.1",
    "torch>=2.8.0",
    "transformers>=4.55.3",
]
